fix(ble/TASK-373): widen profile-loop counters to size_t
Switches the JsonArray-iteration counter in
lib/PedalLogic/src/config_loader.cpp:154 (loadConfig) and
lib/PedalLogic/src/config_loader_merge.cpp:65 (mergeConfig) from
uint8_t to size_t. Both loops compared the narrow counter against
JsonArray::size() (size_t), which CodeQL flagged as
cpp/comparison-with-wider-type — the counter could in principle wrap
to zero and never terminate.

In practice the secondary guard `i < hardwareConfig.numProfiles` (where
numProfiles is uint8_t) bounds the counter at 255, so behaviour is
unchanged. config_loader.cpp adds a static_cast<uint8_t> at the
profileManager.addProfile() call site since the counter is now wider
than the API expects; the cast is safe by construction.

Clears CodeQL alerts #1 (config_loader.cpp) and #2
(config_loader_merge.cpp) so the v0.5.0 release PR #9 can pass the
GHAS code-scanning gate.

+++ lib/PedalLogic/src/config_loader_merge.cpp
@@ -62,7 +62,7 @@ bool ConfigLoader::mergeConfig(ProfileManager& profileManager,
 
     JsonArray profiles = doc["profiles"];
 
-    for (uint8_t newIdx = 0; newIdx < profiles.size() && newIdx < hardwareConfig.numProfiles;
+    for (size_t newIdx = 0; newIdx < profiles.size() && newIdx < hardwareConfig.numProfiles;
          newIdx++)
     {
         JsonObject profileJson = profiles[newIdx];
